Move opensnp analysis script up one level. Create a util.py file.
